Refactor: Remove Gilroy font and add Rubik for Arabic
This commit removes all instances of the Gilroy font and replaces it with the system default.
Additionally, the Rubik font has been added and will be used when the application language is set to Arabic.
The table names for Fortress and SFQ in Arabic localization have been updated to use the English table names.
The Google Play developer link has also been updated.